### 2024-08-23 #7133 @mprib
Thank you for taking this on, @mooniean!

In the past I reviewed a JOSS project ([JeFaPaTo](https://github.com/openjournals/joss-reviews/issues/6425)) with a similar stack of tools used in Caliscope (python/qt/mediapipe). The lead on that project (timozen) may be well suited to comment on this. Otherwise I have no suggestions.
