### 2024-08-27 #7133 @mooniean
Thank you for suggesting, @mprib!

Hi 👋  @timozen @DavidPagnon @MrMallIronmaker - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> #7133

Thanks in advance!
