### 2024-08-28 #7155 @editorialbot
[REVIEW]: Caliscope: GUI Based Multicamera Calibration and Motion Tracking
**Submitting author:** @mprib (<a href="http://orcid.org/0000-0001-9243-468X">Donald Prible</a>)
**Repository:** https://github.com/mprib/caliscope
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.1
