### 2024-09-05 #7155 @davidpagnon
Thank you for making this tool available! I heard good things about it a few months ago but did not find time to test it, so this review is the perfect occasion.

The installation procedure and the execution are quite user-friendly, and the output is the same as AniPose or Pose2Sim so it can be reused as is (or almost, I'll implement very minimal changes to my code so that people can just copy paste your caliscope Config.toml file into their Pose2Sim calibration folder).

Several aspects make it a pain-free first experience:
- A GUI that launches without having to type any Python command
- Clear tutorial and guidelines, and videos at each but the last stage (pose estimation)
- Demonstration data which constitutes a clean model to follow

## Paper:
- It is clear, short, and concise.
- I absolutely don't want to push it as there is obvisouly a conflict of interest, but you could cite the [Pose2Sim calibration](https://github.com/perfanalytics/pose2sim) in your literature. It does not use bundle adjustment, is not as convenient as your abstract class, but it also provide 5 distortion coefficients and allows the user to calibrate with elements of the scene if they forgot to take a calibration board with them.
- More information could be given on the pose estimation part, especially on the holistic opensim optimization, the rig generation, and how to use it with Rigmarole (unless you want to keep it as a bonus feature outside of the scope of Caliscope).

## Doc:
- The text on the GitHub.io website is not the same as in the Readme.md file. As I am not sure which is the most up to date, I suppose it could be confusing to other users as well. Maybe you could have a minimal Readme.md file that redirects to the website?
- I installed it within a conda environment instead of venv, but it still goes smoothly
- You specify that the user either needs to have their cameras already synchronized and with the same number of frames, or to have a frame_time_history.csv file. However, you do not redirect to your multiwebcam package (actually you do in the Readme.md but not in the documentation). It would be very convenient to give this hint to the user.

- Although there can be some indetermination on rotations, most people still use a checkerboard for calibration. Nothing important but it would be nice to include this option!
- The code testing seems to be testing almost all functions and classes, which is great. However, I think the workflow could do some linting (none done for now), test on more OS (only Linux for now), and on more python versions (only 3.11 for now).
  Obviously, according to the instructions you give, you've done some manual testing on windows/MacOS/Linux, but it could be made automatic at each push to the repository. See for example [this package](https://github.com/davidpagnon/Sports2D/blob/main/.github/workflows/continuous-integration.yml), which tests on all OS, with python versions 3.8 to 3.11.
  It's also a bit strange that the installation procedure asks for Python 3.10 while the tests run on 3.11. Out of curiosity, the pyproject.toml requires python >=3.10, <3.12: what prevents it from running on other versions? You make the user install Caliscope via pip, but tests use poetry.
  Really not needed but I take advantage of this point to make a suggestion: you could add another workflow to automatically publish your package to pypi upon each new release (see (this example workflow](https://github.com/davidpagnon/Sports2D/blob/main/.github/workflows/publish-on-release.yml))

## GUI:
- It's a matter of personal taste, but if the terminal window is not wide enough, the ASCII art is collapsed and it almost looks like a bug. I would suggest removing it.
- To make everything clearer, what would you think of adding a description for each button upon hovering over them with the mouse? This is a small adjustment but it could make the user experience slighlty easier.
- The input boxes in which I enter integers are too small, the values are hidden on my screen (e.g.: row x col, camera count, etc.). It would be nice to widen these boxes.
  ![Capture d'écran 2024-09-04 164440](https://github.com/user-attachments/assets/be01d35b-dd99-4488-8ac3-9ac914dd060e)
- I do not quite get how the Charuco board generation works. Once I have set the number of rows and columns and the square size, why are there no margins on both sides of the image when I increase both its width and height?
- It could be useful to add a button to stop camera calibration, volume calibration, and post-processing once they have started. Or just to make these buttons clickable a second time to abort. For example, this can be useful if we process long videos and realize we chose wrong parameters.
- What is the use of "Add grid" and "Calibrate" in the Camera tab? It seems like Autocalibrate is the only useful button.
- "Zoom" only takes effect when "Apply distortion" is ticked.
- It would make more sense to move the "Calibrate capture volume" to the "Capture volume" tab, in the same way that "Camera calibration" is in the "Cameras" tab.
- It is hard to know along which axis to adjust the coordinate system since there is no clear way of knowing which color corresponds to which axis. A simple way would be to simply comply to the xyz -> RGB standard.
- X+, X-, etc do not invert the axis, but rotate the 3D points around axes by 90°. I would simply change the buttons to X+90°, X-90°, etc.
- What is holistic OpenSim? OpenSim is not installed, so how do you perform scaling, inverse kinematics, and rigging? Is it a custom algorithm? I would like to know a bit more about it.

## Side remarks (which would take more time to implement)
As you point it out in the paper, Caliscope's main goal is to make multi-camera calibration easy and straight-forward for non experts. However, you also include a tool to perform pose estimation, triangulate it, and rig the output. I feel like calibration is a huge problem, and having a tool dedicated solely for that purpose could be quite convenient for countless people. Calibration is also needed for the study of animals, objects, or even for architecture, and most people would not care about human pose. And for those who care, MediaPipe/BlazePose is getting outdated so they may want to use something more recent such as [RTMlib](https://github.com/Tau-J/rtmlib). Moreover, that last part seems to be a bit less polished. What would you think of decorrelating these two parts and making them two separate tools?

On the other hand, you coded quite a convenient toolbox to record from multiple webcams and synchronize them. This could be made clearer on the paper that this is available, and actually automatically provides the required "frame_time_history.csv". I do realize this may sound contradictory with my previous point, but I feel like it would be even better to directly include it in the Caliscope. It would not make the toolbox mheavier since the requirements are the same, and people would be able to record and calibrate a scene from one single window. As far as I'm concerned and in a very interested way, I would love to add such a tool to the dependencies of Pose2Sim.

Off-topic: I believe the toolbox used to be called "pyxy3d". I like the new name, which looks much less like a bunch of letters collapsed together and sounds nice and easy to remember.
