### 2024-09-13 mprib/caliscope#636 @mprib
Add Tooltip hints in GUI
From feedback related to JOSS Submission:

> To make everything clearer, what would you think of adding a description for each button upon hovering over them with the mouse? This is a small adjustment but it could make the user experience slightly easier.
