### 2024-09-13 mprib/caliscope#637 @mprib
Add ability to abort calibration 
Feeback from JOSS review:

> It could be useful to add a button to stop camera calibration, volume calibration, and post-processing once they have started. Or just to make these buttons clickable a second time to abort. For example, this can be useful if we process long videos and realize we chose wrong parameters.
