### 2024-09-13 mprib/caliscope#636 @mprib
Make sure to include +90 descriptions in tooltips for 3D axis rotations
