### 2024-09-14 c008783
Update pytest.yml
