+++ .github/workflows/pytest.yml
@@ -18,7 +18,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ['3.11']
+        python-version: ['3.10','3.11']
 
     steps:
     - uses: actions/checkout@v2
