+++ .github/workflows/pytest.yml
@@ -16,9 +16,22 @@ jobs:
       github.event.pull_request.merged == true ||
       github.event_name == 'workflow_dispatch'
 
+    #strategy:
+    #  matrix:
+    #    python-version: ['3.10','3.11']
+
     strategy:
+      fail-fast: false
       matrix:
-        python-version: ['3.10','3.11']
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.10", "3.11"]
+        include:
+          - os: ubuntu-latest
+            path: ~/.cache/pip
+          - os: macos-latest
+            path: ~/Library/Caches/pip
+          - os: windows-latest
+            path: ~\AppData\Local\pip\Cache
 
     steps:
     - uses: actions/checkout@v2
