### 2024-09-14 3608da9
basic updates in response to DavidPagnon's feedback
