+++ docs/post_processing.md
@@ -13,6 +13,14 @@
 5. Click the `Process` button to begin the landmark tracking and triangulation.
 6. 3D landmark positions will be visualized and you can open the subfolder to inspect the landmark tracking on the recordings or to access the trajectory output files
 
+## Tracker Outputs
+
+Current options for the tracker outputs are built on Google's [Mediapipe]() and include pipelines for general [Pose](), [Hands](), and [Face]().
+The [Holistic]() tracker combines all three outputs.
+While the Holistic tracker offers improved tracking of the face and hands compared to the Pose model, the number of points it supplies can quickly become unweildy (several hundred for the face).
+The Holistic OpenSim model filters out many of these points that may be extraneous to users primarily interested in gross skeletal movement. 
+
+
 ## Metarig Generation
 
 For the Holistic OpenSim tracker you can generate a metarig configuration file. This will provide a set of parameters that can scale segments of a skeletal model based on the average distances between various landmarks throughout the motion capture. With a more accurately scaled skeletal model, inverse kinematics can more successfully approximate the true movement.
@@ -24,4 +32,5 @@ For the Holistic OpenSim tracker you can generate a metarig configuration file. 
     - motion blur can substantially compromise landmark recognition
     - using a higher frame rate can reduce motion blur
       - this will require more light to maintain good illumination
+
 
