+++ mkdocs.yml
@@ -34,7 +34,8 @@ nav:
   - Project Setup: 
     - Overview: project_setup.md
     - Example Project: sample_data.md
-  - Workflow:
+  - Workflow: 
+    - Video Capture: video_capture.md
     - Workspace: workspace.md
     - Charuco Board: calibration_board.md
     - Cameras: intrinsic_calibration.md
