### 2024-09-14 e0c6ce0
update bibliography
