+++ paper/paper.bib
@@ -19,11 +19,9 @@
 
 @misc{cinematographydatabaseIndieViconSystem2021,
   title = {Indie {{Vicon System Cost}} and {{Solo Operator Guide}}},
-  author = {Workman, Matthew},
   year = {2021},
   month = nov,
   urldate = {2024-01-11},
-  url = {https://www.youtube.com/watch?v=YgT6XY6ldj8},
   collaborator = {{Cinematography Database}},
   annotation = {YouTube}
 }
@@ -44,17 +42,16 @@
   pmid = {34592148}
 }
 
-@article{lugaresiMediaPipeFrameworkBuilding2019,
+@misc{lugaresiMediaPipeFrameworkBuilding2019,
   title = {{{MediaPipe}}: {{A Framework}} for {{Building Perception Pipelines}}},
   author = {Lugaresi, Camillo and Tang, Jiuqiang and Nash, Hadon and McClanahan, Chris and Uboweja, Esha and Hays, Michael and Zhang, Fan and Chang, Chuo-Ling and Yong, Ming Guang and Lee, Juhyun and Chang, Wan-Teh and Hua, Wei and Georg, Manfred and Grundmann, Matthias},
   year = {2019},
   month = jun,
-  journal = {arXiv preprint arXiv:1906.08172},
   eprint = {1906.08172},
-  eprinttype = {arxiv},
   primaryclass = {cs},
   doi = {10.48550/arXiv.1906.08172},
-  urldate = {2024-01-12}
+  urldate = {2024-01-12},
+  archiveprefix = {arXiv}
 }
 
 @article{mathisDeepLabCutMarkerlessPose2018,
@@ -74,25 +71,10 @@
   langid = {english}
 }
 
-@inproceedings{triggsBundleAdjustmentModern2000,
-  title = {Bundle {{Adjustment}} --- {{A Modern Synthesis}}},
-  booktitle = {Vision {{Algorithms}}: {{Theory}} and {{Practice}}},
-  author = {Triggs, Bill and McLauchlan, Philip F. and Hartley, Richard I. and Fitzgibbon, Andrew W.},
-  editor = {Triggs, Bill and Zisserman, Andrew and Szeliski, Richard},
-  year = {2000},
-  pages = {298--372},
-  publisher = {Springer},
-  address = {Berlin, Heidelberg},
-  doi = {10.1007/3-540-44480-7_21},
-  isbn = {978-3-540-44480-0},
-  langid = {english}
-}
-@article{mayorovLargescaleBundleAdjustment,
-  title = {Large-Scale Bundle Adjustment in SciPy},
+@misc{mayorovLargescaleBundleAdjustment,
+  title = {Large-Scale Bundle Adjustment in Scipy --- {{SciPy Cookbook}} Documentation},
   author = {Mayorov, Nikolay},
   journal = {SciPy Cookbook},
-  year = {2016},
-  month = oct,
   urldate = {2024-01-11},
   howpublished = {https://scipy-cookbook.readthedocs.io/items/bundle\_adjustment.html}
 }
@@ -114,6 +96,35 @@
   volume = {2},
   issn = {2624-9367},
   urldate = {2024-01-21}
+}
+
+@article{pagnonPose2SimOpensourcePython2022,
+  title = {{{Pose2Sim}}: {{An}} Open-Source {{Python}} Package for Multiview Markerless Kinematics},
+  author = {Pagnon, David and Domalain, Mathieu and Reveret, Lionel},
+  year = {2022},
+  month = sep,
+  journal = {Journal of Open Source Software},
+  volume = {7},
+  number = {77},
+  pages = {4362},
+  issn = {2475-9066},
[…]
