### 2024-09-14 mprib/caliscope#637 @mprib
Just commenting here to note that implementing this will likely be a long-term goal and not a short term goal. Experience suggests to me that killing and restarting threads is a very easy way to create very bad memory leaks that are exceptionally difficult to track down.
