### 2024-09-15 ec978ad
First pass at linting changes based on ruff
