+++ caliscope/gui/post_processing_widget.py
@@ -30,14 +30,14 @@ class PostProcessingWidget(QWidget):
 
     def __init__(self, controller:Controller):
         super(PostProcessingWidget, self).__init__()
-        self.controller = controller 
+        self.controller = controller
         self.config = self.controller.config
 
         self.sync_index_cursors = {}  # track where the slider is for each playback...
         self.recording_folders = QListWidget()
         self.update_recording_folders()
 
-        self.vis_widget = PlaybackTriangulationWidget(self.controller.camera_array)  
+        self.vis_widget = PlaybackTriangulationWidget(self.controller.camera_array)
 
         self.tracker_combo = QComboBox()
         self.vizualizer_title = QLabel()
@@ -54,7 +54,7 @@ class PostProcessingWidget(QWidget):
         self.refresh_visualizer() # must happen before placement to create vis_widget and vizualizer_title
         self.place_widgets()
         self.connect_widgets()
-        
+
 
     def set_current_xyz(self):
 
@@ -90,7 +90,7 @@ class PostProcessingWidget(QWidget):
         # add each folder to the QListWidget
         for folder in dir_list:
             self.recording_folders.addItem(folder)
-        
+
         if len(dir_list)>0:
             self.recording_folders.setCurrentRow(0)
 
@@ -112,7 +112,7 @@ class PostProcessingWidget(QWidget):
     @property
     def archived_config_path(self):
         return Path(self.processed_subfolder, "config.toml")
-        
+
 
     @property
     def xy_base_path(self):
@@ -123,13 +123,13 @@ class PostProcessingWidget(QWidget):
     @property
     def active_tracker_enum(self):
         return self.tracker_combo.currentData()
-        
+
     @property
     def metarig_config_path(self):
         file_name = f"metarig_config_{self.tracker_combo.currentData().name}.json"
         result = Path(self.processed_subfolder, file_name)
         return result
-        
+
 
     @property
     def active_folder(self):
@@ -140,15 +140,15 @@ class PostProcessingWidget(QWidget):
             active_folder: str = self.recording_folders.currentItem().text()
         else:
             active_folder: str = self.recording_folders.currentItem().text()
-            
+
         return active_folder
 
     @property
     def active_recording_path(self)-> Path:
         p = Path(self.controller.workspace_guide.recording_dir, self.active_folder)
         logger.info(f"Active recording path is {p}")
-        return p        
-        
+        return p
+
     @property
     def viz_title_html(self):
         if self.xyz_processed_path.exists():
@@ -223,10 +223,10 @@ class PostProcessingWidget(QWidget):
         that can be displayed to the user and will align 
         """
         return Path(self.active_recording_path,self.active_tracker_enum.name, "config.toml")
-        
+
     def refresh_visualizer(self):
         logger.info("Refreshing vizualizer within post_processing widget")
-        
+
         if self.archived_config_path.exists():  # processing has been done and their is a camera array that can be loaded
             stored_config = Configurator(self.archived_config_path.parent)
             presented_camera_array = stored_config.get_camera_array()
@@ -235,7 +235,7 @@ class PostProcessingWidget(QWidget):
 
         self.vis_widget.update_camera_array(presented_camera_array)
 
-        
+
         self.set_current_xyz()
         self.vizualizer_title.setText(self.viz_title_html)
         self.update_enabled_disabled()
@@ -260,7 +260,7 @@ class PostProcessingWidget(QWidget):
         self.vis_widget.slider.setEnabled(True)
 
     def update_enabled_disabled(self):
-        # set availability of metarig generation 
+        # set availability of metarig generation
         logger.info("Checking if metarig config can be created...")
         tracker = self.tracker_combo.currentData().value()
         logger.info(tracker)
[…]
