+++ caliscope/gui/extrinsic_calibration_widget.py
@@ -10,22 +10,16 @@ from enum import Enum
 
 import cv2
 from PySide6.QtCore import Qt, Signal, QThread
-from PySide6.QtGui import QImage, QPixmap, QIcon
+from PySide6.QtGui import QImage, QPixmap
 from PySide6.QtWidgets import (
     QApplication,
-    QSizePolicy,
     QWidget,
     QSpinBox,
     QScrollArea,
-    QComboBox,
-    QCheckBox,
-    QDialog,
     QGroupBox,
-    QDoubleSpinBox,
     QHBoxLayout,
     QLabel,
     QPushButton,
-    QSlider,
     QVBoxLayout,
 )
 
@@ -34,7 +28,6 @@ from caliscope.session.session import LiveSession
 from caliscope.gui.frame_builders.paired_frame_builder import PairedFrameBuilder
 from caliscope.cameras.synchronizer import Synchronizer
 from caliscope import __root__
-from caliscope.gui.navigation_bars import NavigationBarNext
 
 # the boards needed before a pair could be used to bridge pairs without common corners
 MIN_THRESHOLD_FOR_EARLY_CALIBRATE = 5
@@ -49,14 +42,14 @@ class ExtrinsicCalibrationWidget(QWidget):
     calibration_complete = Signal()
     calibration_initiated = Signal()
     terminate = Signal()
-     
+
     def __init__(self,session:LiveSession):
 
         super(ExtrinsicCalibrationWidget, self).__init__()
         self.session = session
         self.synchronizer:Synchronizer = self.session.synchronizer
 
-        logger.info(f"about to check if synchronizer has a sync packet")
+        logger.info("about to check if synchronizer has a sync packet")
         while not hasattr(self.session.synchronizer, "current_sync_packet"):
             logger.info("waiting for synchronizer to create first sync packet")
             sleep(.5)
@@ -69,17 +62,17 @@ class ExtrinsicCalibrationWidget(QWidget):
         self.frame_rate_spin.setValue(self.synchronizer.fps_target)
         self.board_count_spin = QSpinBox()
         self.board_count_spin.setValue(self.paired_frame_builder.board_count_target)
-        
+
         self.stereo_frame_display = QLabel()
-        # self.navigation_bar = NavigationBarNext() 
+        # self.navigation_bar = NavigationBarNext()
         self.possible_action = PossibleActions.CollectData
         self.calibrate_collect_btn = QPushButton(self.possible_action.value)
-        
+
         self.place_widgets()
-        self.connect_widgets()        
+        self.connect_widgets()
 
         self.update_btn_eligibility()
-    
+
     def shutdown_threads(self):
         """
         There may be some lingering threads running when the extrinsic calibrator loses focus
@@ -96,17 +89,17 @@ class ExtrinsicCalibrationWidget(QWidget):
             self.calibrate_collect_btn.setEnabled(True)
         else:
             self.calibrate_collect_btn.setEnabled(False)
-    
+
 
     def place_widgets(self):
         self.setLayout(QVBoxLayout())
-        
+
         self.settings_group = QGroupBox("Settings")
         self.settings_group.setLayout(QHBoxLayout())
         self.settings_group.layout().addWidget(QLabel("Frame Rate:"))
-        self.settings_group.layout().addWidget(self.frame_rate_spin)       
+        self.settings_group.layout().addWidget(self.frame_rate_spin)
         self.settings_group.layout().addWidget(QLabel("Target Board Count:"))
-        self.settings_group.layout().addWidget(self.board_count_spin)       
+        self.settings_group.layout().addWidget(self.board_count_spin)
 
         self.layout().addWidget(self.settings_group)
[…]
