+++ caliscope/gui/frame_emitters/tools.py
@@ -1,5 +1,4 @@
 import cv2
-import numpy as np
 from PySide6.QtGui import QImage
 import caliscope.logger
 

[…]
