+++ caliscope/calibration/stereocalibrator.py
@@ -68,12 +68,8 @@ class StereoCalibrator:
             get_coverage_region, axis=1, args=(self.ports,)
         )
 
-        points_w_pivoted_ports = points_w_pivoted_ports.filter(
-            ["sync_index", "point_id", "coverage_region"]
-        )
-        points_w_regions = point_data.merge(
-            points_w_pivoted_ports, "left", ["sync_index", "point_id"]
-        )
+        points_w_pivoted_ports = points_w_pivoted_ports.filter(["sync_index", "point_id", "coverage_region"])
+        points_w_regions = point_data.merge(points_w_pivoted_ports, "left", ["sync_index", "point_id"])
 
         return points_w_regions
 
@@ -95,9 +91,7 @@ class StereoCalibrator:
 
         # find the primary region each board is in, then merge in total counts
         all_boards = (
-            self.all_point_data.filter(
-                ["port", "sync_index", "point_id", "coverage_region"]
-            )
+            self.all_point_data.filter(["port", "sync_index", "point_id", "coverage_region"])
             .groupby(["port", "sync_index", "coverage_region"])
             .agg({"point_id": "count"})
             .reset_index()
@@ -115,16 +109,12 @@ class StereoCalibrator:
             .reset_index()
             .rename(columns={"coverage_region": "primary_coverage_region"})
             .drop("point_count", axis=1)  # no longer means anthing
-            .merge(
-                board_points, "left", ["port", "sync_index"]
-            )  # merge in a point_count that means something
+            .merge(board_points, "left", ["port", "sync_index"])  # merge in a point_count that means something
         )
 
         return all_boards
 
-    def get_stereopair_data(
-        self, pair: tuple, boards_sampled: int, random_state=1
-    ) -> pd.DataFrame or None:
+    def get_stereopair_data(self, pair: tuple, boards_sampled: int, random_state=1) -> pd.DataFrame or None:
         # convenience function to get the points that are in the overlap regions of the pairs
         def in_pair(row: int, pair: tuple):
             """
@@ -136,15 +126,11 @@ class StereoCalibrator:
 
             a, b = str(a), str(b)
 
-            region_check = ("_" + a + "_") in row.coverage_region and (
-                "_" + b + "_"
-            ) in row.coverage_region
+            region_check = ("_" + a + "_") in row.coverage_region and ("_" + b + "_") in row.coverage_region
             return region_check and port_check
 
         # flag the points that belong to the pair overlap regions
-        self.all_point_data["in_pair"] = self.all_point_data.apply(
-            in_pair, axis=1, args=(pair,)
-        )
+        self.all_point_data["in_pair"] = self.all_point_data.apply(in_pair, axis=1, args=(pair,))
 
         # group points into boards and get the total count for sample weighting below
         pair_points = self.all_point_data[self.all_point_data["in_pair"] == True]
@@ -169,13 +155,9 @@ class StereoCalibrator:
             sample_weight = pair_boards["point_count"] ** 2
 
             # get the randomly selected subset
-            selected_boards = pair_boards.sample(
-                n=sample_size, weights=sample_weight, random_state=random_state
-            )
+            selected_boards = pair_boards.sample(n=sample_size, weights=sample_weight, random_state=random_state)
 
-            selected_pair_points = pair_points.merge(
-                selected_boards, "right", "sync_index"
-            )
+            selected_pair_points = pair_points.merge(selected_boards, "right", "sync_index")
         else:
             logger.info(f"For pair {pair} there are no shared boards")
             selected_pair_points = None
@@ -210,9 +192,7 @@ class StereoCalibrator:
                 self.config[config_key]["translation"] = translation
[…]
