+++ tests/test_export.py
@@ -9,8 +9,9 @@ from caliscope.export import xyz_to_wide_labelled, xyz_to_trc
 import pandas as pd
 
 
-original_data_path = Path( __root__, "tests", "sessions","4_cam_recording","recording_1", "HOLISTIC")
-working_data_path = Path( __root__, "tests", "sessions_copy_delete","4_cam_recording","recording_1", "HOLISTIC")
+original_data_path = Path(__root__, "tests", "sessions", "4_cam_recording", "recording_1", "HOLISTIC")
+working_data_path = Path(__root__, "tests", "sessions_copy_delete", "4_cam_recording", "recording_1", "HOLISTIC")
+
 
 def test_export():
     copy_contents(original_data_path, working_data_path)
@@ -34,10 +35,10 @@ def test_export():
     trc_path = Path(xyz_csv_path.parent, f"{xyz_csv_path.stem}.trc")
     assert not trc_path.exists()
 
-    xyz_to_trc(xyz,tracker,time_history_path, target_path=trc_path)
+    xyz_to_trc(xyz, tracker, time_history_path, target_path=trc_path)
     assert trc_path.exists()
     # %%
 
-if __name__ == "__main__":
 
+if __name__ == "__main__":
     test_export()
