+++ caliscope/trackers/helper.py
@@ -1,6 +1,7 @@
 import cv2
 import numpy as np
 
+
 def apply_rotation(frame: np.ndarray, rotation_count: int) -> np.ndarray:
     if rotation_count == 0:
         pass
@@ -14,12 +15,10 @@ def apply_rotation(frame: np.ndarray, rotation_count: int) -> np.ndarray:
     return frame
 
 
-def unrotate_points(
-    xy: np.ndarray, rotation_count: int, frame_width: int, frame_height: int
-) -> np.ndarray:
+def unrotate_points(xy: np.ndarray, rotation_count: int, frame_width: int, frame_height: int) -> np.ndarray:
     xy_unrotated = xy.copy()
 
-    if rotation_count == 0 or len(xy) ==0:
+    if rotation_count == 0 or len(xy) == 0:
         pass
     elif rotation_count in [1, -3]:
         # Reverse of 90 degrees clockwise rotation

[…]
