+++ .github/workflows/pytest.yml
@@ -17,11 +17,9 @@ jobs:
       github.event_name == 'workflow_dispatch'
 
     strategy:
-      fail-fast: false
       matrix:
         python-version: ['3.10','3.11']
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.10", "3.11"]
         include:
           - os: ubuntu-latest
             path: ~/.cache/pip
