### 2024-09-15 9832eb6
additional unsafe fixes
