### 2024-09-15 63ab5d1
no bare exceptions
