### 2024-09-15 e30454e
removing vestigial widgets
