+++ caliscope/gui/extrinsic_calibration_widget.py
@@ -1,279 +1,0 @@
-import caliscope.logger
-
-logger = caliscope.logger.get(__name__)
-
-import sys
-from pathlib import Path
-from threading import Thread, Event
-from time import sleep, perf_counter
-from enum import Enum
-
-import cv2
-from PySide6.QtCore import Qt, Signal, QThread
-from PySide6.QtGui import QImage, QPixmap
-from PySide6.QtWidgets import (
-    QApplication,
-    QWidget,
-    QSpinBox,
-    QScrollArea,
-    QGroupBox,
-    QHBoxLayout,
-    QLabel,
-    QPushButton,
-    QVBoxLayout,
-)
-
-# Append main repo to top of path to allow import of backend
-from caliscope.session.session import LiveSession
-from caliscope.gui.frame_builders.paired_frame_builder import PairedFrameBuilder
-from caliscope.cameras.synchronizer import Synchronizer
-from caliscope import __root__
-
-# the boards needed before a pair could be used to bridge pairs without common corners
-MIN_THRESHOLD_FOR_EARLY_CALIBRATE = 5
-
-
-class PossibleActions(Enum):
-    CollectData = "Collect Data"
-    Terminate = "Terminate"
-    Calibrate = "Calibrate"
-
-
-class ExtrinsicCalibrationWidget(QWidget):
-    calibration_complete = Signal()
-    calibration_initiated = Signal()
-    terminate = Signal()
-
-    def __init__(self, session: LiveSession):
-        super(ExtrinsicCalibrationWidget, self).__init__()
-        self.session = session
-        self.synchronizer: Synchronizer = self.session.synchronizer
-
-        logger.info("about to check if synchronizer has a sync packet")
-        while not hasattr(self.session.synchronizer, "current_sync_packet"):
-            logger.info("waiting for synchronizer to create first sync packet")
-            sleep(0.5)
-
-        self.paired_frame_builder = PairedFrameBuilder(self.synchronizer, board_count_target=30)
-        self.paired_frame_emitter = PairedFrameEmitter(self.paired_frame_builder)
-        self.paired_frame_emitter.start()
-
-        self.frame_rate_spin = QSpinBox()
-        self.frame_rate_spin.setValue(self.synchronizer.fps_target)
-        self.board_count_spin = QSpinBox()
-        self.board_count_spin.setValue(self.paired_frame_builder.board_count_target)
-
-        self.stereo_frame_display = QLabel()
-        # self.navigation_bar = NavigationBarNext()
-        self.possible_action = PossibleActions.CollectData
-        self.calibrate_collect_btn = QPushButton(self.possible_action.value)
-
-        self.place_widgets()
-        self.connect_widgets()
-
-        self.update_btn_eligibility()
-
-    def shutdown_threads(self):
-        """
-        There may be some lingering threads running when the extrinsic calibrator loses focus
-        This may be causing python to overload and pyqt to segfault during the calibration process
-        if I've moved from the extrinsic calibration widget to a different one...
-        """
-        logger.info("Unsubscribe paired frame builder from sync notice")
-        self.paired_frame_builder.unsubscribe_from_synchronizer()
-        logger.info("signal paired frame emitter to stop collecting frames")
-        self.paired_frame_emitter.keep_collecting.clear()
-
-    def update_btn_eligibility(self):
-        if self.session.is_extrinsic_calibration_eligible():
-            self.calibrate_collect_btn.setEnabled(True)
-        else:
-            self.calibrate_collect_btn.setEnabled(False)
-
-    def place_widgets(self):
-        self.setLayout(QVBoxLayout())
-
-        self.settings_group = QGroupBox("Settings")
-        self.settings_group.setLayout(QHBoxLayout())
-        self.settings_group.layout().addWidget(QLabel("Frame Rate:"))
[…]
