+++ tests/test_no_video_calibrate.py
@@ -2,6 +2,7 @@ import caliscope.logger
 
 logger = caliscope.logger.get(__name__)
 
+
 def test_no_video_calibrate():
     """
     Saving out of video may be maxing system resources when high res/many cameras.
