+++ caliscope/post_processing/gap_filling.py
@@ -1,5 +1,5 @@
-import pandas as pd
 import numpy as np
+import pandas as pd
 
 import caliscope.logger
 
