### 2024-09-15 1e26a12
fix filter of array
