### 2024-09-15 598ce34
remove vestigial widgets
