+++ caliscope/gui/old_main_widget.py
@@ -1,303 +1,0 @@
-import sys
-from pathlib import Path
-from threading import Thread
-
-import rtoml
-from PySide6.QtCore import Qt
-from PySide6.QtGui import QAction, QIcon
-from PySide6.QtWidgets import (
-    QApplication,
-    QDockWidget,
-    QFileDialog,
-    QMainWindow,
-    QMenu,
-    QWidget,
-)
-
-import caliscope.logger
-from caliscope import __root__, __settings_path__
-from caliscope.configurator import Configurator
-from caliscope.gui.charuco_widget import CharucoWidget
-from caliscope.gui.extrinsic_calibration_widget import ExtrinsicCalibrationWidget
-from caliscope.gui.live_camera_config.intrinsic_calibration_widget import (
-    IntrinsicCalibrationWidget,
-)
-from caliscope.gui.log_widget import LogWidget
-from caliscope.gui.post_processing_widget import PostProcessingWidget
-from caliscope.gui.recording_widget import RecordingWidget
-from caliscope.gui.vizualize.calibration.capture_volume_widget import CaptureVolumeWidget
-from caliscope.session.session import LiveSession, SessionMode
-
-logger = caliscope.logger.get(__name__)
-
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        super(MainWindow, self).__init__()
-
-        self.app_settings = rtoml.load(__settings_path__)
-
-        self.setWindowTitle("Caliscope")
-        self.setWindowIcon(QIcon(str(Path(__root__, "caliscope/gui/icons/pyxy_logo.svg"))))
-        self.setMinimumSize(500, 500)
-
-        # File Menu
-        self.menu = self.menuBar()
-
-        # CREATE FILE MENU
-        self.file_menu = self.menu.addMenu("&File")
-        self.open_project_action = QAction("New/Open Project", self)
-        self.file_menu.addAction(self.open_project_action)
-
-        # Open Recent
-        self.open_recent_project_submenu = QMenu("Recent Projects...", self)
-        # Populate the submenu with recent project paths;
-        # reverse so that last one appended is at the top of the list
-        for project_path in reversed(self.app_settings["recent_projects"]):
-            self.add_to_recent_project(project_path)
-
-        self.file_menu.addMenu(self.open_recent_project_submenu)
-
-        self.exit_pyxy3d_action = QAction("Exit", self)
-        self.file_menu.addAction(self.exit_pyxy3d_action)
-
-        # CREATE CAMERA MENU
-        self.cameras_menu = self.menu.addMenu("&Cameras")
-        self.connect_cameras_action = QAction("Co&nnect Cameras", self)
-        self.cameras_menu.addAction(self.connect_cameras_action)
-        self.connect_cameras_action.setEnabled(False)
-
-        self.disconnect_cameras_action = QAction("&Disconnect Cameras", self)
-        self.cameras_menu.addAction(self.disconnect_cameras_action)
-        self.disconnect_cameras_action.setEnabled(False)
-
-        # CREATE MODE MENU
-        self.mode_menu = self.menu.addMenu("&Mode")
-        self.charuco_mode_select = QAction(SessionMode.Charuco.value)
-        self.intrinsic_mode_select = QAction(SessionMode.IntrinsicCalibration.value)
-        self.extrinsic_mode_select = QAction(SessionMode.ExtrinsicCalibration.value)
-        self.capture_volume_mode_select = QAction(SessionMode.CaptureVolumeOrigin.value)
-        self.recording_mode_select = QAction(SessionMode.Recording.value)
-        self.triangulate_mode_select = QAction(SessionMode.Triangulate.value)
[…]
