+++ caliscope/recording/video_recorder.py
@@ -192,19 +192,6 @@ class VideoRecorder:
         self.destination_folder = destination_folder
         # create the folder if it doesn't already exist
         self.destination_folder.mkdir(exist_ok=True, parents=True)
-
-        # # Because calibration files are nested in a calibration directory, need to go
-        # # to parent.parent to get the config.toml file
-        # if self.destination_folder.parent.stem == "calibration":
-        #     source_config_path = Path(self.destination_folder.parent.parent, "config.toml")
-        # else:   # just a regular recording
-        #     source_config_path = Path(self.destination_folder.parent, "config.toml")
-
-        # No longer storing config file with recordings....can't know when they were done relative to calibration so will only complicate things..
-        # source_config_path = find_config_file(self.destination_folder)
-        # duplicate_config_path = Path(self.destination_folder,"config.toml")
-        # shutil.copy2(source_config_path,duplicate_config_path)
-
         self.recording = True
         self.recording_thread = Thread(
             target=self.save_data_worker,
