+++ caliscope/cameras/synchronizer.py
@@ -125,7 +125,7 @@ class Synchronizer:
             self.port_frame_count[port] += 1
 
             logger.debug(
-                f"Frame data harvested from reel {frame_packet.port} with index {frame_index} and frame time of {frame_packet.frame_time}"
+                f"Frame data harvested from reel {frame_packet.port} with index {frame_index} and frame time of {frame_packet.frame_time}"  # noqa E501
             )
 
         logger.info(f"Frame harvester for port {port} completed")
@@ -259,7 +259,7 @@ class Synchronizer:
                 q.put(self.current_sync_packet)
                 if self.current_sync_packet is not None:
                     logger.debug(
-                        f"Placing new synched frames packet on queue with {self.current_sync_packet.frame_packet_count} frames"
+                        f"Placing new synched frames packet on queue with {self.current_sync_packet.frame_packet_count} frames"  # noqa E501
                     )
                     logger.debug(f"Placing new synched frames with index {self.current_sync_packet.sync_index}")
 
