+++ caliscope/cameras/camera.py
@@ -56,10 +56,10 @@ class Camera:
             self.port = port
             self.capture = test_capture
             self.active_port = True
+
             # limit buffer size so that you are always reading the latest frame
-            self.capture.set(
-                cv2.CAP_PROP_BUFFERSIZE, 1
-            )  # from https://stackoverflow.com/questions/58293187/opencv-real-time-streaming-video-capture-is-slow-how-to-drop-frames-or-getanother thread signaled a change to mediapipe overley-sync
+            # from https://stackoverflow.com/questions/58293187/opencv-real-time-streaming-video-capture-is-slow-how-to-drop-frames-or-getanother thread signaled a change to mediapipe overley-sync  # noqa: E501
+            self.capture.set(cv2.CAP_PROP_BUFFERSIZE, 1)
 
             self.ignore = False  # flag camera during single camera setup to be ignored in the future
 
