### 2024-09-15 d8a6e66
revise steps to use pip and ruff
