+++ .github/workflows/pytest.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }} 
     if: >
       github.event.pull_request.merged == true ||
       github.event_name == 'workflow_dispatch'
@@ -21,7 +21,6 @@ jobs:
       matrix:
         python-version: ['3.10','3.11']
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.10", "3.11"]
         include:
           - os: ubuntu-latest
             path: ~/.cache/pip
@@ -31,40 +30,22 @@ jobs:
             path: ~\AppData\Local\pip\Cache
 
     steps:
-    - uses: actions/checkout@v2
+    - name: Checkout Code
+      uses: actions/checkout@v4
 
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
 
-    # note this install is likely relevant for all ubuntu users...
-    # taken from https://www.reddit.com/r/github/comments/10aenmk/help_linux_action_cant_find_opengl_all_of_a_sudden/
-    - name: Update packages
-      run: sudo apt-get update
-
-    - name: Install packages
-      run: sudo apt-get install --fix-missing libgl1-mesa-dev
-
-    # - name: Install X11 and xcb dependencies
-    #   run: |
-    #     sudo apt-get update
-    #     sudo apt-get install -y xvfb libxcb1 x11-apps
-      
-    - name: Install Poetry
-      run: |
-        curl -sSL https://install.python-poetry.org | python3 -
-
-    - name: Configure Poetry
-      run: |
-        echo "$HOME/.local/bin" >> $GITHUB_PATH
-        poetry config virtualenvs.create false
-
     - name: Install dependencies
-      run: poetry install
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install ruff pytest
+        pip install -e .
 
-    # - name: Run tests with xvfb
-    #   run: xvfb-run --auto-servernum --server-args='-screen 0, 1024x768x24' poetry run pytest
-
-    - name: Run tests 
-      run: poetry run pytest
+    - name: Lint with Ruff
+      run: ruff check .
+    - name: Test with pytest
+      run: pytest
+        
