### 2024-09-15 06b3289
refine os specific pre test steps
