### 2024-09-15 f7d2c35
set macOS env variables
