### 2024-09-15 6113b38
adjust for documented off-by-one error that is now Issue #640
