### 2024-09-15 mprib/caliscope#641 @mprib
Scale Input Boxes to Improve visability of numbers
Report from DavidPagnon that the spin boxes were not large enough to display text:

![364946408-be01d35b-dd99-4488-8ac3-9ac914dd060e](https://github.com/user-attachments/assets/265a6f95-f54e-4662-a92d-ad5dd0f747f7)

Documenting target appearance of GUI:

![image](https://github.com/user-attachments/assets/cfb477be-bb11-427b-b06c-784ba55378f7)

Also, for reference, this is how the charuco should update when paper size changes:

![image](https://github.com/user-attachments/assets/6b52ea87-2300-4434-be50-abca47824622)
