+++ .github/workflows/pytest.yml
@@ -17,7 +17,7 @@ jobs:
       github.event_name == 'workflow_dispatch'
 
     strategy:
-      fail-fast: false
+      fail-fast: true
       matrix:
         python-version: ['3.10','3.11']
         os: [ubuntu-latest, macos-latest, windows-latest]
