### 2024-09-16 c981205
remove dark theme as a dependency
