+++ .github/workflows/pytest.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        python-version: ['3.10','3.11']
+        python-version: ['3.10','3.11', '3.12']
         os: [ubuntu-latest, macos-latest, windows-latest]
         include:
           - os: ubuntu-latest
