### 2024-09-16 7bdd1b4
Incompatibility with mediapipe + macOS + python 3.12
