### 2024-09-16 mprib/caliscope#644 @mprib
Move Calibrate Capture Volume button from main tab to Capture Volume Tab
From JOSS review GUI feedback:

>     * It would make more sense to move the "Calibrate capture volume" to the "Capture volume" tab, in the same way that "Camera calibration" is in the "Cameras" tab.
