### 2024-09-16 #7155 @mprib
@davidpagnon ,

Thank you again for your comments. My replies are woven in below. For suggested GUI improvements that I have not yet been able to resolve, I have opened up issues on the repo and provide links to those here.

Regards,

Mac

> Thank you for making this tool available! I heard good things about it a few months ago but did not find time to test it, so this review is the perfect occasion.
>
> Several aspects make it a pain-free first experience:
>
>    A GUI that launches without having to type any Python command
>
>    Clear tutorial and guidelines, and videos at each but the last stage (pose estimation)
>
>    Demonstration data which constitutes a clean model to follow
>
>
> ## Paper:
>
>  It is clear, short, and concise.
>
>  I absolutely don't want to push it as there is obvisouly a conflict of interest, but you could cite the [Pose2Sim calibration](https://github.com/perfanalytics/pose2sim) in your literature. It does not use bundle adjustment, is not as convenient as your abstract class, but it also provide 5 distortion coefficients and allows the user to calibrate with elements of the scene if they forgot to take a calibration board with them.

Thank you for the suggestion and I absolutely agree that it's worthwhile to discuss the two projects. I have included this new paragraph in the paper to highlight similarities and distinctions:

*Pose2Sim (Pagnon et al., 2022) is another tool that merges both camera calibration and markerless pose estimation. Similar to Caliscope, Pose2Sim employs a camera model composed of 4 pinhole camera parameters and 5 distortion parameters. Additionally, both projects export output to the .trc file format to facilitate integration with the biomechanical modelling software OpenSim. While Pose2Sim does not perform bundle adjustment to refine the extrinsic camera estimates, it does provide a number of features that are valuable components of a motion tracking workflow which are not present in Caliscope and would be useful in future motion tracking tools. These include the ability to calibrate camera extrinsics based on scene elements, the capacity to distinguish between multiple subjects in view at once, and more sophisticated triangulation methods that can incorporate the confidence of a model’s prediction of a pose landmark.*

>
>   More information could be given on the pose estimation part, especially on the holistic opensim optimization, the rig generation, and how to use it with Rigmarole (unless you want to keep it as a bonus feature outside of the scope of Caliscope).

I have added text to the Post Processing section of the docs to clarify the Holistic OpenSim tracker.

Essentially, it is a filtered subset of the Mediapipe Holistic output. In test usage, files of the raw Holistic output became unweildy due to containing several hundred points for the face and dozens for the hands; but then using the smaller output of Mediapipe Pose became problematic because it converts to a much lower resolution for point detection. Holistic Opensim was my attempt to make a smaller file with better detection that could track the wrist and skull better.

Regarding the rig generation/Rigmarole: the "bonus features" characterization is spot on. My intention is to try to keep Caliscope largely agnostic to these downstream tasks and to primarily try to do "one thing well" (camera calibration),: and to provide the foundation of a framework for integrating future pose estimation tools into the pipeline.

>
> ## Doc:
>
>  The text on the GitHub.io website is not the same as in the Readme.md file. As I am not sure which is the most up to date, I suppose it could be confusing to other users as well. Maybe you could have a minimal Readme.md file that redirects to the website?

I have added the following to the ReadMe for clarity:

*This README provides a general overview and quick guide to install `Calicope` on your system. For a more detailed description of the calibration process and workflow, please see our [docs](https://mprib.github.io/caliscope/).*

>    I installed it within a conda environment instead of venv, but it still goes smoothly
>
>    You specify that the user either needs to have their cameras already synchronized and with the same number of frames, or to generate a frame_time_history.csv file. However, you do not redirect to your multiwebcam package (actually you do in the Readme.md but not in the documentation). It would be very convenient to give this hint to the user.

Thank you for the feedback. I have added a Video Capture tab to the workflow section of the docs that directs to MultiWebCam.
>
>    Although there can be some indetermination on rotations, most people still use a checkerboard for calibration. Nothing important but it would be nice to include this option.

Thank you. I would need to better understand how to resolve the indeterminate rotations to know how this might be implemented but it does seem like an interesting future addition.

>
>    The code testing seems to be testing almost all functions and classes, which is great. However, the workflow could do some linting (none done for now), test on more OS (only Linux for now), and with more Python versions (only 3.11 for now).

Using your CI action linked below as a template I've updated the testing to python 3.10, and 3.11 across macOS/Linux/Windows. And finally got around to making sure the code lints with ruff.

>   Obviously, according to your instructions, you've done some manual testing on Windows/MacOS/Linux, but it could be made automatic upon each push to the repository. See for example [this package](https://github.com/davidpagnon/Sports2D/blob/main/.github/workflows/continuous-integration.yml), which tests on all OS, with python versions 3.8 to 3.11.

See comment above. Thanks for the useful resource.

>   It's also a bit strange that the installation procedure asks for Python 3.10 while the tests run on 3.11. Out of curiosity, the pyproject.toml requires python >=3.10, <3.12: what prevents it from running on other versions? Lastly, you make the user install Caliscope via pip, but tests use poetry instead.

For simplicity I now have the tests installing with pip. I have expanded the tests to include 3.10 and 3.11.

There was a patch of python 3.9 that came out that I found was associated with random segfaults in PyQt, so that one got taken off the list.

There are currently issues with the combination of Python 3.12 + Mediapipe + MacOS, so 3.12 is not included. This should not effect the core camera calibration workflow, so future developments that do not involve mediapipe (or macOS) should not experience this roadblock.

>    Really not needed, but I take advantage of this point to make a suggestion: you could add another workflow to automatically publish your package to pypi upon each new release (see (this example workflow](https://github.com/davidpagnon/Sports2D/blob/main/.github/workflows/publish-on-release.yml))

But typing `poetry publish` really does give me so much joy!

>
> ## GUI:
>
>    It's a matter of personal taste, but if the terminal window is not wide enough, the ASCII art collapses and it almost looks like a bug. I would suggest removing it.

Removed

>    To make everything clearer, what would you think of adding a description for each button upon hovering over them with the mouse? This is a small adjustment but it could make the user experience slightly easier.

Issue created to add tooltips: https://github.com/mprib/caliscope/issues/636

>    The input boxes in which I enter integers are too small. It would be nice to widen these boxes.
>    ![Capture d'écran 2024-09-04 164440](https://private-user-images.githubusercontent.com/54667644/364946408-be01d35b-dd99-4488-8ac3-9ac914dd060e.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjY1MDE0NzQsIm5iZiI6MTcyNjUwMTE3NCwicGF0aCI6Ii81NDY2NzY0NC8zNjQ5NDY0MDgtYmUwMWQzNWItZGQ5OS00NDg4LThhYzMtOWFjOTE0ZGQwNjBlLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MTYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTE2VDE1MzkzNFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTY4ZDkwNjUyNGIwODM2M2ZlNjAyMjEwNmFhMTkxNmU0OTJlMDBiNDdjOTNlZDNiMzQzZjc5MGQzNDFjYmRjYTMmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.hT1zGDxiqI2qr7hKp6wPo8Ylowfi2qvCKia3P9aIQ3w)

Issue created to widen text boxes: https://github.com/mprib/caliscope/issues/641

In the issue above I include some screenshots of how it is intended to appear (and shows on my system). I have recently removed a qt theme for simplicity's sake which may also impact the character legibility.

>   I do not quite get how the Charuco board generation works. Once I have set the number of rows and columns and the square size, why are there no margins on both sides of the image when I increase both its width and height?

I'm wondering if images are getting clipped in the GUI? The issue linked above (641) includes some screenshots of how the board should display. My hope is that at the very least the properly rendered board is what is save out as a `png`.

>    It could be useful to add a button to stop camera calibration, volume calibration, and post-processing once they have started. Or just to make these buttons clickable a second time to abort. For example, this can be useful if we process long videos and realize we chose wrong parameters.

Agree this would be an improvement and have added it as an issue, though this is one of those long-term issues that may take some time to reflect on how to best implement : https://github.com/mprib/caliscope/issues/637

Just for context, I have found that killing and re-launching these threads can easily result in pretty bad memory leaks. Realistically, it may take some time before I could implement that improvement in a stable way, though I agree it would be a good future addition.

>    What is the use of "Add grid" and "Calibrate" in the Camera tab? It seems like "Autocalibrate" is the only useful button.

The first two are the original interface and Autocalibrate was added later. I think most of the time autocalibrate will work fine, but I see the value in having the ability to select specific frames as wanted.

>   "Zoom" only takes effect when "Apply distortion" is ticked.

This is as intended. It's possible for a bad distortion model to appear reasonable because it also appears zoomed in when "undistorted". When zoomed out, the bizarre distortion around the image edges is apparent.

>   It would make more sense to move the "Calibrate capture volume" to the "Capture volume" tab, in the same way that "Camera calibration" is in the "Cameras" tab.

I agree that this suggestion would improve the flow. I have created an issue to incorporate this in a future version: https://github.com/mprib/caliscope/issues/644

>    It is hard to know along which axis to adjust the coordinate system since there is no clear way of knowing which color corresponds to which axis. A simple way would be to simply comply to the xyz -> RGB standard.

I definitely acknowledge that step requires experimentation to orient to the controls. The colors are the defaults offered by PyQtGraph's AxisItem object and I do not see a way in their documentation to modify the color selection.

PyQtGraph is amazing, though. Strongly recommend.

>    X+, X-, etc do not invert the axis, but rotate the 3D points around axes by 90°. I would simply change the buttons to X+90°, X-90°, etc.

I will plan to include the X+90, etc in the tooltips for clarity.

>    What is holistic OpenSim? OpenSim is not installed, so how do you perform scaling, inverse kinematics, and rigging? Is it a custom algorithm? I would like to know a bit more about it.

The holistic opensim tracker is discussed above. It is essentially a subset of the holistic mediapipe output that could drive IK on a full skeletal model in OpenSim, but without lots of points for the hands and face.

The scaling is just taking the mean distance between select landmarks across a dynamic calibration trial. Those distances are then used to scale a blender rig that is animated via IK by the subsequent motion trials.

The rigging component is in a separate package, Rigmarole which is based on Blender. Those are definitely side-quests and not the primary goal of the package.

> ## Side remarks (which would take more time to implement)
>
> As you point out in the paper, Caliscope's main goal is to make multi-camera calibration easy and straightforward for non-experts. However, you also include a tool to perform pose estimation, triangulate it, and rig the output. I feel like calibration is a huge problem, and having a tool dedicated solely to that purpose could be quite convenient for countless people. Calibration is also needed for the study of animals, objects, or even for architecture, and most people would not care about human pose. And for those who care, MediaPipe/BlazePose is getting outdated so they may want to use something more recent such as [RTMlib](https://github.com/Tau-J/rtmlib). Moreover, that last bit of the package seems to be a bit less polished. What would you think of decorrelating these two parts and making them two separate tools?

I definitely view the "Post Processing" tab as a proof of concept to demonstrate the quality of the calibration and what could be done with it. I am highly drawn to the idea of "do one thing" and see the value of just having the camera system calibration spit out a config which can be used by by other tools.

Landmark tracking, triangulation, filtering, and inverse kinematics are all such rich domains themselves.
It would be great to have loosely coupled tools that could pass the ball down field to other packages via standardized human readable file formats.

> On the other hand, you coded quite a convenient toolbox to record from multiple webcams and synchronize them. This could be mentioned in the paper, especially since it actually automatically generates the required "frame_time_history.csv" file that is required. I do realize this may sound contradictory to my previous point, but I feel like it would be even better to directly include it in the Caliscope. It would not make the toolbox heavier since the requirements are the same, and people would be able to record and calibrate a scene from one single window.

With video recorded from a synchronized camera system (like FLIR), `frame_time_history.csv` is not required. When that file is not present, frames will just be bundled up in lock-step.

Regarding `MultiWebCam`, I have become hesitant to promote the use of webcams strongly.
Working with them across multiple operating systems is a bit of a nightmare, and my expectation is that bundling `MultiWebCam` with `Caliscope` would create issues.

`MultiWebCam` did allow me to test a lot of things out cheaply while iterating in my home office, but the resolution/frame rate/motion blur will limit many use cases.
That being said, I also know how much the Vicon systems in our lab cost, and my gut tells me that given how good and inexpensive modern phone cameras are, there has to be a better way than the major mocap vendors.

*here I begin a bit of a ramble about the project broadly*

I'm excited to pursue alternate hardware approaches in the future such as FLIR, or more DIY approaches such as raspberry pi's with MIPI cameras.
I strongly suspect we are approaching at a time where inexpensive off-the-shelf components can capture high quality synchronized footage appropriate for research and clinical applications.

I see this project fundamentally as managing the conversion of raw videos to a human readable camera calibration `config.toml`.
I want it to be as agnostic as possible to the adjacent workflows so that it can be easily adapted to integrate with these tools as they emerge and develop.

I'm glad that this review has given an opportunity to connect with someone who is working on these problems as well and hope that we might continue a conversation about them.
I see a future where there is a broad ecosystem of open source tools that can be integrated into a comprehensive motion capture/biomechanical workflow.
My hope with `Caliscope` is to contribute one small link in the chain.

> Off-topic: I believe the toolbox used to be called "pyxy3d". I like the new name, which looks much less like a bunch of letters collapsed together and sounds nice and easy to remember.

Thank you! Naming things is agonizingly difficult...
