### 2024-09-17 mprib/caliscope#641 @davidpagnon
Hi @mprib,

Thank you for taking into account all these suggestions!

I'm still not sure I understand:
Let's say you set the board to have 4 rows x 3 columns. Let's say you specified that the square size is 5.4 cm. Then the Charuco board should be 21.6 x 16.2 cm².
If you specified that the board size should be 25 x 20 cm², shouldn't you have margins on all sides? More specifically, 1.7 cm on top and bottom, and 1.9 on the right and left sides?

![image](https://github.com/user-attachments/assets/46218913-d04d-498b-acca-9cdc8ba26885)
