### 2024-09-17 mprib/caliscope#641 @davidpagnon
Actually, I think I got it: if we did not manage to print the board properly, this can be fixed before proceeding to the next stage by correcting the actual square edge length. Is this it?
