### 2024-09-17 #7155 @davidpagnon
Thank you for this extensive response! I'm satisfied with the answers I did not address.

> Essentially, it is a filtered subset of the Mediapipe Holistic output. In test usage, files of the raw Holistic output became unweildy due to containing several hundred points for the face and dozens for the hands; but then using the smaller output of Mediapipe Pose became problematic because it converts to a much lower resolution for point detection. Holistic Opensim was my attempt to make a smaller file with better detection that could track the wrist and skull better.

So I'm not sure I totally get it. I understand that you "scale a blender rig that is animated via IK" but then you do not use OpenSim scaling and IK , do you?

> Thank you for the feedback. I have added a Video Capture tab to the workflow section of the docs that directs to MultiWebCam.

I feel like it would be worth linking to MultiWebCam at each mention of frame_time_history.csv, or at least [there](https://mprib.github.io/caliscope/project_setup/#frame_time_historycsv:~:text=a%20file%20called-,frame_time_history.csv,-within%20the%20folder) as when I first had a look at the overview section, I thought you had to create your own script or create the file manually, which can be scary for new users.

> But typing poetry publish really does give me so much joy!

Okay, thank you, I don't know much about poetry and did not realize there was such a nice way to publish a package. And more importantly, I would not want to deprive you of such a joy!

> In the issue above I include some screenshots of how it is intended to appear (and shows on my system). I have recently removed a qt theme for simplicity's sake which may also impact the character legibility.

I answered on this issue: https://github.com/mprib/caliscope/issues/641

> Just for context, I have found that killing and re-launching these threads can easily result in pretty bad memory leaks. Realistically, it may take some time before I could implement that improvement in a stable way, though I agree it would be a good future addition.

Okay, good to know!

> > What is the use of "Add grid" and "Calibrate" in the Camera tab? It seems like "Autocalibrate" is the only useful button.
>
> The first two are the original interface and Autocalibrate was added later. I think most of the time autocalibrate will work fine, but I see the value in having the ability to select specific frames as wanted.

So the use of "Add grid" and "Calibrate" is for selecting specific frames? I do not see anything about it in the doc, it may be something to add?

> Working with them across multiple operating systems is a bit of a nightmare, and my expectation is that bundling MultiWebCam with Caliscope would create issues.

Okay, it makes sense!

What are FLIR cameras? I thought they were mostly for infrared video capture, are you telling me they are also good for RGB capture, synchronization, and that they are rather cheap? In any case, I quite like the idea of using Raspberry Pis with MIPI cameras!

> I see a future where there is a broad ecosystem of open source tools that can be integrated into a comprehensive motion capture/biomechanical workflow.

Well, I would not be against the idea of a community creating the Blender of Mocap 😄 It would be quite a long-term project but it does not seem totally out of reach indeed!
