### 2024-09-17 mprib/caliscope#641 @mprib
> this can be fixed before proceeding to the next stage by correcting the actual square edge length. Is this it?

Yes, it is this exactly. The scale of the world is set by the actual size of the squares. In practice there is generally a small deviation of the final printed page from the "perfect" intended measurement.

This number ultimately just goes into the charuco board definition in OpenCV: https://docs.opencv.org/3.4/d0/d3c/classcv_1_1aruco_1_1CharucoBoard.html
