### 2024-09-17 657eec0
improve documentation based on JOSS feedback
