### 2024-09-17 824538e
more renaming of HOLISTIC_OPENSIM to SIMPLE_HOLISTIC
