+++ caliscope/trackers/tracker_enum.py
@@ -4,14 +4,14 @@ from caliscope.trackers.charuco_tracker import CharucoTracker
 from caliscope.trackers.face_tracker import FaceTracker
 from caliscope.trackers.hand_tracker import HandTracker
 from caliscope.trackers.holistic.holistic_tracker import HolisticTracker
-from caliscope.trackers.holistic_opensim_tracker import HolisticOpenSimTracker
 from caliscope.trackers.pose_tracker import PoseTracker
+from caliscope.trackers.simple_holistic_tracker import SimpleHolisticTracker
 
 
 class TrackerEnum(Enum):
     HAND = HandTracker
     POSE = PoseTracker
-    HOLISTIC_OPENSIM = HolisticOpenSimTracker
+    SIMPLE_HOLISTIC = SimpleHolisticTracker
     HOLISTIC = HolisticTracker
     CHARUCO = CharucoTracker
     FACE = FaceTracker
