+++ caliscope/trackers/holistic_opensim_tracker.py
@@ -1,325 +1,0 @@
-from queue import Queue
-from threading import Thread
-
-import cv2
-import mediapipe as mp
-import numpy as np
-
-import caliscope.logger
-from caliscope.packets import PointPacket
-from caliscope.tracker import Tracker
-from caliscope.trackers.helper import apply_rotation, unrotate_points
-
-logger = caliscope.logger.get(__name__)
-
-MIN_DETECTION_CONFIDENCE = 0.5
-MIN_TRACKING_CONFIDENCE = 0.95
-
-# The following are from base Pose and can be ignored in favor of
-# better estimated Holistic points
-DRAW_IGNORE_LIST = [
-    "nose",
-    "left_eye_inner",
-    "left_eye",
-    "left_eye_outer",
-    "right_eye_inner",
-    "right_eye",
-    "right_eye_outer",
-    "left_ear",
-    "right_ear",
-    "mouth_left",
-    "mouth_right",
-    "left_wrist_pose",
-    "right_wrist_pose",
-    "left_pinky",
-    "right_pinky",
-    "left_index",
-    "right_index",
-    "left_thumb",
-    "right_thumb",
-]
-
-POINT_NAMES = {
-    # 0: "nose",
-    # 1: "left_eye_inner",
-    # 2: "left_eye",
-    # 3: "left_eye_outer",
-    # 4: "right_eye_inner",
-    # 5: "right_eye",
-    # 6: "right_eye_outer",
-    # 7: "left_ear",
-    # 8: "right_ear",
-    # 9: "mouth_left",
-    # 10: "mouth_right",
-    11: "left_shoulder",
-    12: "right_shoulder",
-    13: "left_elbow",
-    14: "right_elbow",
-    # 15: "left_wrist_pose",
-    # 16: "right_wrist_pose",
-    # 17: "left_pinky",
-    # 18: "right_pinky",
-    # 19: "left_index",
-    # 20: "right_index",
-    # 21: "left_thumb",
-    # 22: "right_thumb",
-    23: "left_hip",
-    24: "right_hip",
-    25: "left_knee",
-    26: "right_knee",
-    27: "left_ankle",
-    28: "right_ankle",
-    29: "left_heel",
-    30: "right_heel",
-    31: "left_foot_index",
-    32: "right_foot_index",
-    100: "right_wrist",
-    101: "right_thumb_CMC",
-    102: "right_thumb_MCP",
-    103: "right_thumb_IP",
-    104: "right_thumb_tip",
-    105: "right_index_finger_MCP",
-    106: "right_index_finger_PIP",
-    107: "right_index_finger_DIP",
-    108: "right_index_finger_tip",
-    109: "right_middle_finger_MCP",
-    110: "right_middle_finger_PIP",
-    111: "right_middle_finger_DIP",
-    112: "right_middle_finger_tip",
-    113: "right_ring_finger_MCP",
-    114: "right_ring_finger_PIP",
-    115: "right_ring_finger_DIP",
-    116: "right_ring_finger_tip",
-    117: "right_pinky_MCP",
-    118: "right_pinky_PIP",
-    119: "right_pinky_DIP",
-    120: "right_pinky_tip",
-    200: "left_wrist",
-    201: "left_thumb_CMC",
-    202: "left_thumb_MCP",
-    203: "left_thumb_IP",
-    204: "left_thumb_tip",
-    205: "left_index_finger_MCP",
-    206: "left_index_finger_PIP",
-    207: "left_index_finger_DIP",
-    208: "left_index_finger_tip",
-    209: "left_middle_finger_MCP",
-    210: "left_middle_finger_PIP",
-    211: "left_middle_finger_DIP",
[…]
