### 2024-09-18 #7155 @mprib
@davidpagnon,

I've cropped the conversation for brevity and hope that this addresses any unresolved points. If there are any outstanding questions remaining, please just let me know!

Best,

Mac

---

> > Essentially, it is a filtered subset of the Mediapipe Holistic output. In test usage, files of the raw Holistic output became unweildy due to containing several hundred points for the face and dozens for the hands; but then using the smaller output of Mediapipe Pose became problematic because it converts to a much lower resolution for point detection. Holistic Opensim was my attempt to make a smaller file with better detection that could track the wrist and skull better.
>
> So I'm not sure I totally get it. I understand that you "scale a blender rig that is animated via IK" but then you do not use OpenSim scaling and IK , do you?

That is correct. I had originally created that paired down version with the intention of some ad hoc experimentations in OpenSim and ended up repurposing it for use in Blender where I did some automated tools in a different package.

Given the opportunity for confusion, I have renamed it to `Simple Holistic`.

> > Thank you for the feedback. I have added a Video Capture tab to the workflow section of the docs that directs to MultiWebCam.
>
> I feel like it would be worth linking to MultiWebCam at each mention of frame_time_history.csv, or at least [there](https://mprib.github.io/caliscope/project_setup/#frame_time_historycsv:~:text=a%20file%20called-,frame_time_history.csv,-within%20the%20folder) as when I first had a look at the overview section, I thought you had to create your own script or create the file manually, which can be scary for new users.

Excellent suggestion. Link to sample data and MultiWebCam added there.

>
> > > What is the use of "Add grid" and "Calibrate" in the Camera tab? It seems like "Autocalibrate" is the only useful button.
> >
> >
> > The first two are the original interface and Autocalibrate was added later. I think most of the time autocalibrate will work fine, but I see the value in having the ability to select specific frames as wanted.
>
> So the use of "Add grid" and "Calibrate" is for selecting specific frames? I do not see anything about it in the doc, it may be something to add?
>

Thank you. Added instructions associated with this to the Intrinsic Calibration docs.

>
> What are FLIR cameras? I thought they were mostly for infrared video capture, are you telling me they are also good for RGB capture, synchronization, and that they are rather cheap? In any case, I quite like the idea of using Raspberry Pis with MIPI cameras!
>

I was sloppy with my language. [Teledyne FLIR](https://www.flir.com/support-center/iis/machine-vision/application-note/configuring-synchronized-capture-with-multiple-cameras/) provides cameras that can record synchronized RGB video (among other types) using a shared hardware trigger.

Aside: a potential advantage that I see for a raspberry pi based camera system is that each "camera server" could provide both image capture and potentially some local landmark estimation. Processing times from a multicamera system can become unwieldy when all frames are processed on a single machine, but if distributed across multiple devices it could be far more manageable.
