### 2024-09-20 #7155 @Timozen
## Installation

I have evaluated the software on MacOS 15 using a MacBook with an M1 processor.

Following the installation procedure in https://mprib.github.io/caliscope/installation/#__tabbed_2_2, the program did not start.

After checking the main [[readme.md](http://readme.md/)](http://readme.md) of the repo, I saw that for MacOS, some env variables need to be set. This resolved the issue ^^

As @davidpagnon already mentioned, the [[readme.md](http://readme.md/)](http://readme.md) and documentation seem to need to be in complete sync; this should be resolved.

## Usage

I will first say that this is the first time I have worked with multi-view pose estimation software. However, I have my fair share of camera calibration experience.

I followed the Example Project page. Please note that I found this 404 link while checking the Overview page: https://mprib.github.io/caliscope/sample_data.md

The tutorial YouTube video of the project https://youtu.be/voE3IKYtuIQ should be, IMHO, more prominent for users so that inexperienced people (like me, in this case, ^^) can follow along.

This was extremely helpful and might make the overall experience smooth!!!

The intrinsic calibration ran quickly on my machine, and the log helped me understand what the tool was trying to do in the background.

The extrinsic calibration ran at the same speed for me.

After the reloading of the workflow, the window size (which I adjusted) reset to the original, which is a minor thing, but it confused me for a second ^^

Initially, I was a bit confused about the repetitive reloading of the workspace; however, I quickly realized that whenever a menu option was unavailable, such as "Post Processing," the underlying reason was that something was not yet in the project folder.

So, I quickly learned to utilize the "reload workspace" button whenever something was added to the project or one processing step finished.

After that, I was able to reproduce the test project results.

When I attempted to extract the "FACE" option, the whole tool crashed without any notice and what the underlying cause is/was.

I could repeat this crash several times without a problem, and I checked the log files, but they contained no actual clue.

```
INFO| caliscope.recording.video_recorder| 101|  End of sync packets signaled...breaking record loop
INFO| caliscope.recording.video_recorder| 149|  releasing video writers...
INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 1
INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 2
INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 3
INFO| caliscope.recording.video_recorder| 156|  Initiate storing of frame history
INFO| caliscope.recording.video_recorder| 176|  Storing frame history to /test/recordings/sitting/FACE/frame_time_history.csv
INFO| caliscope.recording.video_recorder| 159|  Initiate storing of point history
INFO| caliscope.recording.video_recorder| 170|  Storing point data in /test/recordings/sitting/FACE/xy_FACE.csv
INFO| caliscope.recording.video_recorder| 164|  About to emit `all frames saved` signal
INFO| caliscope.post_processing.post_processor|  73|  (Stage 1 of 2): 99% of frames processed for (x,y) landmark detection
INFO| caliscope.post_processing.post_processor|  96|  Reading in (x,y) data..
INFO| caliscope.post_processing.post_processor|  99|  Filling small gaps in (x,y) data
INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 1. Filling gaps that are 3 frames or less...
INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 2. Filling gaps that are 3 frames or less...
INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 3. Filling gaps that are 3 frames or less...
INFO| caliscope.post_processing.gap_filling|  71|  (x,y) gap filling complete
INFO| caliscope.post_processing.post_processor| 101|  Beginning data triangulation
INFO| caliscope.cameras.camera_array| 209|  Creating camera array projection matrices
INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 1
INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 2
INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 3
INFO| caliscope.triangulate.triangulation| 201|  Assembling undistorted dataframe
INFO| caliscope.triangulate.triangulation| 120|  About to begin triangulation...due to jit, first round of calculations may take a moment.
[1]    92789 bus error  caliscope
```

However, the tool still wrote the results into the appropriate folder, and reloading the project also worked.
This is not a significant issue, but perhaps you can find a way to capture such crashes and write the log.
Please also note that after restarting the tool, the previous log is gone, so I could not investigate the first crash.
IMHO, this is not a reason for me not to continue the paper's publication, but it should be something you should be aware of.

## GUI

- First of all, I know the effort that goes into creating such a sophisticated GUI; props to that!
- The concerns I would have raised already have been brought up by @davidpagnon and answered by @mprib.
- It would still be helpful if this information is in the tool docs with an explanation.

## Paper

I like the style of the paper. It was written clearly, showing the shortcomings or issues of the current state and which value Caliscope adds to the area.

As most information is also inside the repository and documentation, further explanations are optional from my side.

---
@mprib I still need to check the 'Installation instructions' checkbox, but it would be nice if you could update the online documentation to contain the MacOS problems and fixes.
After that, I will be okay with everything.
