### 2024-09-25 mprib/caliscope#648 @mprib
FACE tool crashing on sample project
> When I attempted to extract the "FACE" option, the whole tool crashed without any notice and what the underlying cause is/was.
>
> I could repeat this crash several times without a problem, and I checked the log files, but they contained no actual clue.
>
> ```
> INFO| caliscope.recording.video_recorder| 101|  End of sync packets signaled...breaking record loop
> INFO| caliscope.recording.video_recorder| 149|  releasing video writers...
> INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 1
> INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 2
> INFO| caliscope.recording.video_recorder| 151|  releasing video writer for port 3
> INFO| caliscope.recording.video_recorder| 156|  Initiate storing of frame history
> INFO| caliscope.recording.video_recorder| 176|  Storing frame history to /test/recordings/sitting/FACE/frame_time_history.csv
> INFO| caliscope.recording.video_recorder| 159|  Initiate storing of point history
> INFO| caliscope.recording.video_recorder| 170|  Storing point data in /test/recordings/sitting/FACE/xy_FACE.csv
> INFO| caliscope.recording.video_recorder| 164|  About to emit `all frames saved` signal
> INFO| caliscope.post_processing.post_processor|  73|  (Stage 1 of 2): 99% of frames processed for (x,y) landmark detection
> INFO| caliscope.post_processing.post_processor|  96|  Reading in (x,y) data..
> INFO| caliscope.post_processing.post_processor|  99|  Filling small gaps in (x,y) data
> INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 1. Filling gaps that are 3 frames or less...
> INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 2. Filling gaps that are 3 frames or less...
> INFO| caliscope.post_processing.gap_filling|  42|  Gap filling for (x,y) data from port 3. Filling gaps that are 3 frames or less...
> INFO| caliscope.post_processing.gap_filling|  71|  (x,y) gap filling complete
> INFO| caliscope.post_processing.post_processor| 101|  Beginning data triangulation
> INFO| caliscope.cameras.camera_array| 209|  Creating camera array projection matrices
> INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 1
> INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 2
> INFO| caliscope.triangulate.triangulation| 193|  Processing points from camera 3
> INFO| caliscope.triangulate.triangulation| 201|  Assembling undistorted dataframe
> INFO| caliscope.triangulate.triangulation| 120|  About to begin triangulation...due to jit, first round of calculations may take a moment.
> [1]    92789 bus error  caliscope
> ```
>
> However, the tool still wrote the results into the appropriate folder, and reloading the project also worked. This is not a significant issue, but perhaps you can find a way to capture such crashes and write the log. Please also note that after restarting the tool, the previous log is gone, so I could not investigate the first crash. IMHO, this is not a reason for me not to continue the paper's publication, but it should be something you should be aware of.
