### 2024-09-25 737fa46
Additional info for MacOS install
