### 2024-09-25 1d9f426
FACE tracker crashing on sample data
