+++ docs/sample_project.md
@@ -1,7 +1,6 @@
 # Sample Project
 
-A complete walkthrough of the calibration and post-processing workflow is shown below. To determine how `caliscope` runs on your hardware, you can download the sample dataset used in this video [here](https://utexas-my.sharepoint.com/:f:/g/personal/priblede_my_utexas_edu/EpJRde1NTPNHpTGrVmPLZA0B_kWTm08KI2QuDjZNY8BV3A?e=tGSwk3). 
-
+A complete walkthrough of the calibration and post-processing workflow is shown below. To determine how `caliscope` runs on your hardware, you can download the sample dataset used in this video [here](https://utexas-my.sharepoint.com/:f:/g/personal/priblede_my_utexas_edu/EpJRde1NTPNHpTGrVmPLZA0B_kWTm08KI2QuDjZNY8BV3A?e=tGSwk3).
 
 Please note that this demo project is meant to illustrate the workflow as simply as possible.There are a variety of improvements that could be made to improve the final results, including:
 
@@ -11,5 +10,4 @@ Please note that this demo project is meant to illustrate the workflow as simply
 - larger calibration board
 - hardware synchronization rather than folding together frames based on timestamps
 
-
- <iframe width="560" height = "315" src="https://www.youtube.com/embed/voE3IKYtuIQ?si=U-ivFqX0trbjG5QA" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+<iframe width="560" height = "315" src="https://www.youtube.com/embed/voE3IKYtuIQ?si=U-ivFqX0trbjG5QA" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
