+++ caliscope/trackers/tracker_enum.py
@@ -1,11 +1,13 @@
 from enum import Enum
 
 from caliscope.trackers.charuco_tracker import CharucoTracker
-from caliscope.trackers.face_tracker import FaceTracker
 from caliscope.trackers.hand_tracker import HandTracker
 from caliscope.trackers.holistic.holistic_tracker import HolisticTracker
 from caliscope.trackers.pose_tracker import PoseTracker
 from caliscope.trackers.simple_holistic_tracker import SimpleHolisticTracker
+
+# Temporarily removed face tracker because crashing sample project
+# from caliscope.trackers.face_tracker import FaceTracker
 
 
 class TrackerEnum(Enum):
