+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "caliscope"
-version = "0.5.0"
+version = "0.5.1"
 description = "GUI based multicamera calibration that integrates with 2D landmark tracking to triangulate 3D landmark positions"
 authors = ["Mac Prible <prible@gmail.com>"]
 license = "BSD-2-Clause"
