### 2024-09-26 #7155 @Timozen
@mprib, that all sounds very good to me! I checked the last missing box in my review panel, and I would be good to go now to continue with the next part of the submission phase @mooniean :)

Concerning the opened issue https://github.com/mprib/caliscope/issues/648, if you need my support and testing again, just let me know over there ^^
