### 2024-10-08 99ed678
Remove original template
