+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
 affiliations:
  - name: The University of Texas at Austin, United States
    index: 1
-date: 31 January 2024
+date: 08 October 2024
 bibliography: paper.bib
 
 ---
