### 2024-10-08 #7155 @mprib
@mooniean,

The JOSS release has been uploaded to Zenodo: https://zenodo.org/records/13905848

DOI: 10.5281/zenodo.13905848

I have double checked the ORCID/Author/Title/License info.

Please let me know if there is anything that I need to clean up on my end.

Thank you!

Mac
