+++ paper/paper.bib
@@ -18,12 +18,13 @@
 }
 
 @misc{cinematographydatabaseIndieViconSystem2021,
+  author = {{Cinematography Database}},
   title = {Indie {{Vicon System Cost}} and {{Solo Operator Guide}}},
   year = {2021},
   month = nov,
   urldate = {2024-01-11},
-  collaborator = {{Cinematography Database}},
-  annotation = {YouTube}
+  url = {https://www.youtube.com/watch?v=YgT6XY6ldj8},
+  note = {YouTube video}
 }
 
 @article{karashchukAniposeToolkitRobust2021,
@@ -74,6 +75,7 @@
 @misc{mayorovLargescaleBundleAdjustment,
   title = {Large-Scale Bundle Adjustment in Scipy --- {{SciPy Cookbook}} Documentation},
   author = {Mayorov, Nikolay},
+  year = {n.d.},
   journal = {SciPy Cookbook},
   urldate = {2024-01-11},
   howpublished = {https://scipy-cookbook.readthedocs.io/items/bundle\_adjustment.html}
