### 2024-10-09 bdb9539
update bib
