+++ paper/paper.bib
@@ -73,12 +73,12 @@
 }
 
 @misc{mayorovLargescaleBundleAdjustment,
-  title = {Large-Scale Bundle Adjustment in Scipy --- {{SciPy Cookbook}} Documentation},
+  title = {Large-Scale Bundle Adjustment in Scipy --- SciPy Cookbook Documentation},
   author = {Mayorov, Nikolay},
-  year = {n.d.},
+  year = {2015},
   journal = {SciPy Cookbook},
   urldate = {2024-01-11},
-  howpublished = {https://scipy-cookbook.readthedocs.io/items/bundle\_adjustment.html}
+  url = {https://scipy-cookbook.readthedocs.io/items/bundle\_adjustment.html}
 }
 
 @misc{mmposecontributorsOpenMMLabPoseEstimation2020,
