+++ paper/paper.bib
@@ -73,7 +73,7 @@
 }
 
 @misc{mayorovLargescaleBundleAdjustment,
-  title = {Large-Scale Bundle Adjustment in Scipy --- SciPy Cookbook Documentation},
+  title = {Large-Scale Bundle Adjustment in Scipy},
   author = {Mayorov, Nikolay},
   year = {2015},
   journal = {SciPy Cookbook},
