+++ paper/paper.bib
@@ -78,7 +78,7 @@
   year = {2015},
   journal = {SciPy Cookbook},
   urldate = {2024-01-11},
-  url = {https://scipy-cookbook.readthedocs.io/items/bundle\_adjustment.html}
+  url = {https://scipy-cookbook.readthedocs.io/items/bundle_adjustment.html}
 }
 
 @misc{mmposecontributorsOpenMMLabPoseEstimation2020,
