### 2024-10-10 #7155 @mprib
Upon reviewing the proof I have made minor updates to the paper.bib to address some issues.
