### 2024-10-11 #7155 @mooniean
@mprib check the missing DOIs (and try to find the ones for the "SKIP" too)
