+++ paper/paper.bib
@@ -97,7 +97,8 @@
   journal = {Frontiers in Sports and Active Living},
   volume = {2},
   issn = {2624-9367},
-  urldate = {2024-01-21}
+  urldate = {2024-01-21},
+  doi = { 10.3389/fspor.2020.00050}
 }
 
 @article{pagnonPose2SimOpensourcePython2022,
