### 2024-10-13 #7155 @mprib
@mooniean,

I have updated `paper.md` with the edit and `paper.bib with the missing DOI. I have double checked the SKIPs and it appears that these do not have DOIs.

The version number in the repo, the zenodo release, and PyPI is v0.5.1.

Please let me know if anything additional is needed.

Thank you!

Mac
