### 2024-10-14 #7155 @crvernon
:wave: @mprib - Thanks for producing a really nice submission.  I just need you to address the following before I accept this for publication:

- [ ] LINE 45:  There should be a space in between "OpenCV" and your citation.

No need to conduct a new release.  Just let me know when this has been fixed and I'll accept.

Thanks!
