+++ paper/paper.md
@@ -39,7 +39,7 @@ Anipose [@karashchukAniposeToolkitRobust2021], which integrates with DeepLabCut,
 To facilitate the calibration process, Anipose simplifies the underlying camera intrinsic parameters, using one parameter each for the pinhole camera and distortion.
 A process of iterative bundle adjustment is used to remove the impact of outlier results in the calibration.
 
-Caliscope carries forward the triangulation code of Anipose, but provides dedicated estimation of camera intrinsics according to the standard camera model employed by OpenCV[@bradskyOpenCVLibrary2000]. It therefore estimates 4 parameters for the pinhole camera model, and 5 parameters for the lens distortion. This is in contrast to the 2 parameters estimated by Anipose (focal length and a single distortion paramters). With more accurate and precise intrinsic parameters, the bundle adjustment computes quickly without the need for an iterative approach.
+Caliscope carries forward the triangulation code of Anipose, but provides dedicated estimation of camera intrinsics according to the standard camera model employed by OpenCV [@bradskyOpenCVLibrary2000]. It therefore estimates 4 parameters for the pinhole camera model, and 5 parameters for the lens distortion. This is in contrast to the 2 parameters estimated by Anipose (focal length and a single distortion paramters). With more accurate and precise intrinsic parameters, the bundle adjustment computes quickly without the need for an iterative approach.
 
 FreeMocap [@cherianFreeMoCapFreeOpen2024] is another notable open-source motion capture project. It uses the original calibration process developed by Anipose. Unlike Anipose, FreeMocap uses Google's Mediapipe [@lugaresiMediaPipeFrameworkBuilding2019] for pose tracking, which avoids the more complicated setup required by DeepLabCut. Caliscope also uses this approach for ease of execution, though defines the `Tracker` as an abstract base class that is implemented with multiple versions of Mediapipe tracking (i.e. pose, hands, and holistic). This flexible framework is intendeded to facilitate the integration of additional landmark tracking tools in the future, such as MMPose [@mmposecontributorsOpenMMLabPoseEstimation2020].
 
